requirements.txt
pytz
numpy
pandas
requests
six
urllib3
pytest
setuptools>=28.8.0
websocket-client
goto-statement